// app/model/competitie.stand.ts
/**
 * Created by Jan Bart Local on 7-7-2016.
 */
export interface ICompStand {
    splId: string;
    splNaam: string;
    tsCar: number;
    tsGemid: number;
    rangPunt: number;
    rangRend: number;
    punten: number;
    rendement: number;
    aantWed: number;
    totCar: number;
    totBrt: number;
    totGemid: number;
    hgSerie: number;
    hgGemid: number;
}