// app/model/speler.stand.ts
/**
 * Created by Jan Bart Local on 10-7-2016.
 */
export interface ISpelerStand {
    tegId: string;
    tegNaam: string;
    tegTsCar: number;
    tegTsGemid: number;
    wedDatum: string;
    splWedCar: number;
    splWedBrt: number;
    splWedSer: number;
    splWedGem: number;
    splWedRend: number;
    splWedPunt: number;
    tegWedCar: number;
    tegWedSer: number;
}